+++ src/index.js
@@ -2,7 +2,16 @@
  * Re-export all functions defined in individual files.
  */
 
-export { dailyAverage } from "./dailyAverage.js";
-export { diurnalAverage } from "./diurnalAverage.js";
+export { dailyStats } from "./dailyStats.js";
+export { diurnalStats } from "./diurnalStats.js";
 export { pm_nowcast } from "./nowcast.js";
 export { trimDate } from "./trimDate.js";
+export {
+  arrayMin,
+  arrayMax,
+  arrayCount,
+  arraySum,
+  arrayMean,
+  roundAndUseNull,
+  useNull,
+} from "./utils.js";