+++ NEWS.md
@@ -8,5 +8,5 @@
 Initial release includes the following algorithms:
 
 - `dailyAverage()` -- Creates daily averages for the specified time zone.
-- `diurnalAverage()` -- Creates hour-of-day averages for the specified time zone.\* `nowcast()` -- Returns an array of NowCast values derived from the incoming time series.
+- `diurnalStats()` -- Creates hour-of-day averages for the specified time zone.\* `nowcast()` -- Returns an array of NowCast values derived from the incoming time series.
 - `trimDate()` -- Trims time series data to complete days in the specified time zone.